RUMM-2239 CR feedback - rename
