+++ Sources/Datadog/Core/DD/ConsoleLogger.swift
@@ -1,0 +1,98 @@
+/*
+ * Unless explicitly stated otherwise all files in this repository are licensed under the Apache License Version 2.0.
+ * This product includes software developed at Datadog (https://www.datadoghq.com/).
+ * Copyright 2019-2020 Datadog, Inc.
+ */
+
+import Foundation
+
+/// The `CoreLogger` printing to debugger console.
+internal struct ConsoleLogger: CoreLogger {
+    /// The prefix applied to all core logs.
+    private static let prefix = "[DATADOG SDK] 🐶 → "
+
+    /// The date provider for annotating core logs.
+    private let dateProvider: DateProvider
+    /// Formatter used to format the time accordingly for local device.
+    private let dateFormatter: DateFormatterType
+    /// The print function.
+    private let printFunction: (String) -> Void
+    /// V1's verbosity level. Only logs above or equal to this level wil be printed.
+    private let currentVerbosityLevel: () -> LogLevel?
+
+    init(
+        dateProvider: DateProvider,
+        timeZone: TimeZone,
+        printFunction: @escaping (String) -> Void ,
+        verbosityLevel: @escaping () -> LogLevel?
+    ) {
+        self.dateProvider = dateProvider
+        self.dateFormatter = presentationDateFormatter(withTimeZone: timeZone)
+        self.printFunction = printFunction
+        self.currentVerbosityLevel = verbosityLevel
+    }
+
+    func debug(_ message: @autoclosure () -> String, error: Error? = nil) {
+        guard let verbosityLevel = currentVerbosityLevel(), LogLevel.debug.rawValue >= verbosityLevel.rawValue else {
+            return // if no `Datadog.verbosityLevel` is set or it is set above this level
+        }
+
+        print(message: message(), error: error, emoji: "")
+    }
+
+    func warn(_ message: @autoclosure () -> String, error: Error? = nil) {
+        guard let verbosityLevel = currentVerbosityLevel(), LogLevel.warn.rawValue >= verbosityLevel.rawValue else {
+            return // if no `Datadog.verbosityLevel` is set or it is set above this level
+        }
+
+        print(message: message(), error: error, emoji: "⚠️")
+    }
+
+    func error(_ message: @autoclosure () -> String, error: Error? = nil) {
+        guard let verbosityLevel = currentVerbosityLevel(), LogLevel.error.rawValue >= verbosityLevel.rawValue else {
+            return // if no `Datadog.verbosityLevel` is set or it is set above this level
+        }
+
+        print(message: message(), error: error, emoji: "🔥")
+    }
+
+    func critical(_ message: @autoclosure () -> String, error: Error? = nil) {
+        guard let verbosityLevel = currentVerbosityLevel(), LogLevel.critical.rawValue >= verbosityLevel.rawValue else {
+            return // if no `Datadog.verbosityLevel` is set or it is set above this level
+        }
+
+        print(message: message(), error: error, emoji: "⛔️")
+    }
+
+    // MARK: - Private
+
+    private func print(message: @autoclosure () -> String, error: Error?, emoji: String) {
+        var log = buildMessageString(message: message(), emoji: emoji)
+
+        if let error = error {
+            log += "\n\nError details:\n\(buildErrorString(error: error))"
+        }
+
+        printFunction(log)
+    }
+
+    private func buildMessageString(message: @autoclosure () -> String, emoji: String) -> String {
+        let prefix = ConsoleLogger.prefix
+        let time = dateFormatter.string(from: dateProvider.currentDate())
+
+        if !emoji.isEmpty {
+            return "\(prefix)\(time) \(emoji) \(message())"
+        } else {
+            return "\(prefix)\(time) \(message())"
+        }
+    }
+
+    private func buildErrorString(error: Error) -> String {
+        let dderror = DDError(error: error)
+        return """
+        → type: \(dderror.type)
+        → message: \(dderror.message)
+        → stack: \(dderror.stack)
+        """
+    }
+}
