+++ Sources/Datadog/Core/DD/CoreLogger.swift
@@ -6,93 +6,97 @@
 
 import Foundation
 
-/// The `CoreLogger` provides methods to collect debug information and detect execution errors of the Datadog SDK.
-internal struct CoreLogger: CoreLoggerType {
-    /// The prefix applied to all core logs.
-    private static let prefix = "[DATADOG SDK] 🐶 → "
+/// The `CoreLogger` protocol defines methods to print debug information and execution errors from Datadog SDK to user console.
+///
+/// It is meant for debugging purposes when using the SDK, hence **it should log information useful and actionable
+/// to the SDK user**. Think of possible logs that we may want to receive from our users when asking them to enable
+/// SDK verbosity and send us their console log.
+internal protocol CoreLogger {
+    /// Print debug message which is meant to self-diagnose possible issues with the SDK.
+    /// It should be used to log all events which might be important for us in diagnosing the SDK
+    /// in user apps (e.g.: printing the SDK version or important aspects of configuration).
+    ///
+    /// No emoji prefix is added by the logger when priting this log to the console.
+    ///
+    /// - Parameters:
+    ///   - message: the message
+    ///   - error: eventual `Error` which will be printed in nice format
+    func debug(_ message: @autoclosure () -> String, error: Error?)
 
-    /// The date provider for annotating core logs.
-    private let dateProvider: DateProvider
-    /// Formatter used to format the time accordingly for local device.
-    private let dateFormatter: DateFormatterType
-    /// The print function.
-    private let printFunction: (String) -> Void
-    /// V1's verbosity level. Only logs above or equal to this level wil be printed.
-    private let currentVerbosityLevel: () -> LogLevel?
+    /// Print error message which indicates **an user error when using the SDK**. It should be used for
+    /// indicating errors that are caused by user fault (e.g. wrong configuration).
+    ///
+    /// The "⚠️" emoji prefix is added by the logger when priting this log to the console.
+    ///
+    /// - Parameters:
+    ///   - message: the message
+    ///   - error: eventual `Error` which will be printed in nice format
+    func warn(_ message: @autoclosure () -> String, error: Error?)
 
-    init(
-        dateProvider: DateProvider,
-        timeZone: TimeZone,
-        printFunction: @escaping (String) -> Void ,
-        verbosityLevel: @escaping () -> LogLevel?
-    ) {
-        self.dateProvider = dateProvider
-        self.dateFormatter = presentationDateFormatter(withTimeZone: timeZone)
-        self.printFunction = printFunction
-        self.currentVerbosityLevel = verbosityLevel
+    /// Print error message which indicates **an error in the SDK**. It shuld be only used for errors
+    /// which are not caused by the user (e.g. due to wrong configuration).
+    ///
+    /// The "🔥" emoji prefix is added by the logger when priting this log to the console.
+    ///
+    /// - Parameters:
+    ///   - message: the message
+    ///   - error: eventual `Error` which will be printed in nice format
+    func error(_ message: @autoclosure () -> String, error: Error?)
+
+    /// Print error message which indicates an error which **makes some part of the SDK unfunctional**.
+    /// It can be used to indicate either fatal SDK errors or user fault.
+    ///
+    /// The "⛔️" emoji prefix is added by the logger when priting this log to the console.
+    ///
+    /// - Parameters:
+    ///   - message: the message
+    ///   - error: eventual `Error` which will be printed in nice format
+    func critical(_ message: @autoclosure () -> String, error: Error?)
+}
+
+extension CoreLogger {
+    /// Print debug message which is meant to self-diagnose possible issues with the SDK.
+    /// It should be used to log all events which might be important for us in diagnosing the SDK
+    /// in user apps (e.g.: printing the SDK version or important aspects of configuration).
+    ///
+    /// No emoji prefix is added by the logger when priting this log to the console.
+    ///
+    /// - Parameters:
+    ///   - message: the message
+    func debug(_ message: @autoclosure () -> String) {
+        debug(message(), error: nil)
     }
 
-    func debug(_ message: @autoclosure () -> String, error: Error? = nil) {
-        guard let verbosityLevel = currentVerbosityLevel(), LogLevel.debug.rawValue >= verbosityLevel.rawValue else {
-            return // if no `Datadog.verbosityLevel` is set or it is set above this level
-        }
-
-        print(message: message(), error: error, emoji: "")
+    /// Print error message which indicates **an user error when using the SDK**. It should be used for
+    /// indicating errors that are caused by user fault (e.g. wrong configuration).
+    ///
+    /// The "⚠️" emoji prefix is added by the logger when priting this log to the console.
+    ///
+    /// - Parameters:
+    ///   - message: the message
+    func warn(_ message: @autoclosure () -> String) {
+        warn(message(), error: nil)
     }
 
-    func warn(_ message: @autoclosure () -> String, error: Error? = nil) {
-        guard let verbosityLevel = currentVerbosityLevel(), LogLevel.warn.rawValue >= verbosityLevel.rawValue else {
-            return // if no `Datadog.verbosityLevel` is set or it is set above this level
-        }
-
-        print(message: message(), error: error, emoji: "⚠️")
+    /// Print error message which indicates **an error in the SDK**. It shuld be only used for errors
+    /// which are not caused by the user (e.g. due to wrong configuration).
+    ///
+    /// The "🔥" emoji prefix is added by the logger when priting this log to the console.
+    ///
+    /// - Parameters:
+    ///   - message: the message
+    func error(_ message: @autoclosure () -> String) {
+        error(message(), error: nil)
     }
 
-    func error(_ message: @autoclosure () -> String, error: Error? = nil) {
-        guard let verbosityLevel = currentVerbosityLevel(), LogLevel.error.rawValue >= verbosityLevel.rawValue else {
-            return // if no `Datadog.verbosityLevel` is set or it is set above this level
-        }
-
-        print(message: message(), error: error, emoji: "🔥")
-    }
-
-    func critical(_ message: @autoclosure () -> String, error: Error? = nil) {
-        guard let verbosityLevel = currentVerbosityLevel(), LogLevel.critical.rawValue >= verbosityLevel.rawValue else {
-            return // if no `Datadog.verbosityLevel` is set or it is set above this level
-        }
-
-        print(message: message(), error: error, emoji: "⛔️")
-    }
-
-    // MARK: - Private
-
-    private func print(message: @autoclosure () -> String, error: Error?, emoji: String) {
-        var log = buildMessageString(message: message(), emoji: emoji)
-
-        if let error = error {
-            log += "\n\nError details:\n\(buildErrorString(error: error))"
-        }
-
-        printFunction(log)
-    }
-
-    private func buildMessageString(message: @autoclosure () -> String, emoji: String) -> String {
-        let prefix = CoreLogger.prefix
-        let time = dateFormatter.string(from: dateProvider.currentDate())
-
-        if !emoji.isEmpty {
-            return "\(prefix)\(time) \(emoji) \(message())"
-        } else {
-            return "\(prefix)\(time) \(message())"
-        }
-    }
-
-    private func buildErrorString(error: Error) -> String {
-        let dderror = DDError(error: error)
-        return """
-        → type: \(dderror.type)
-        → message: \(dderror.message)
-        → stack: \(dderror.stack)
-        """
+    /// Print error message which indicates an error which **makes some part of the SDK unfunctional**.
+    /// It can be used to indicate either fatal SDK errors or user fault.
+    ///
+    /// The "⛔️" emoji prefix is added by the logger when priting this log to the console.
+    ///
+    /// - Parameters:
+    ///   - message: the message
+    func critical(_ message: @autoclosure () -> String) {
+        critical(message(), error: nil)
     }
 }
